chore: move interactive types to separate file
