+++ packages/components/src/components/hds/interactive/index.ts
@@ -5,26 +5,7 @@
 
 import Component from '@glimmer/component';
 import { action } from '@ember/object';
-
-export interface HdsInteractiveArgs {
-  href?: string;
-  isHrefExternal?: boolean;
-  isRouteExternal?: boolean;
-  route?: string;
-  models?: Array<string | number>;
-  model?: string | number;
-  query?: Record<string, string>;
-  'current-when'?: string;
-  replace?: boolean;
-}
-
-export interface HdsInteractiveSignature {
-  Args: HdsInteractiveArgs;
-  Blocks: {
-    default: [];
-  };
-  Element: HTMLAnchorElement | HTMLButtonElement;
-}
+import type { HdsInteractiveSignature } from './types';
 
 export default class HdsInteractiveIndexComponent extends Component<HdsInteractiveSignature> {
   /**
