+++ packages/components/src/components/hds/link/types.ts
@@ -1,4 +1,4 @@
-import type { HdsInteractiveArgs } from '../interactive';
+import type { HdsInteractiveArgs } from '../interactive/types.ts';
 
 export enum HdsLinkIconPositionValues {
   Leading = 'leading',
